Support CiOptron::syncTo if there is no GPS module
With no external GPS module, after setting GPS from TheSkyX,
we'll run into GPS_RECEIVING_VALID_DATA, so a sync on star
is never possible if one doesn't have the optional GPS module.

I had clear skies to test the mount for tonight, and it was in
"outer space" on first boot, so I tried an image link/sync
prior to doing my initial TPoint model.

This fix allowed me to do that and start on the model.